Move windows code to where it belongs, but keep variable in volumecontrol so you can always set it
